more cleanup and String -> Long for temp types

+++ src/main/kotlin/no/kartverket/matrikkel/bygning/routes/v1/BygningRoutes.kt
@@ -40,7 +40,7 @@ fun Route.bygningRouting() {
                     }
                 }
 
-                val bygning = bygningService.getBygning(bygningId, gyldigFraDate)
+                val bygning = bygningService.getBygning(bygningId.toLong(), gyldigFraDate)
 
                 if (bygning != null) {
                     call.respond(bygning)
